fix: address Copilot review comments on Salesforce skill PR
- Return null from captureOAuthResponse when no secrets captured
- Make OAuth auto-capture opt-in (empty capture fields when omitted)
- Fix docstring mismatch in salesforce_query.py authenticate()
- Constrain SF_DOMAIN docs to valid prefix values only
- Remove misleading profile file config section from reference docs
- Fix no-op assertion (or True) in salesforce skill tests
- Bind bearer tokens to exact hostname instead of naive base domain
- Break circular dependency: extract shared helpers to gateway-http-utils.ts
