+++ tests/salesforce-skill.test.ts
@@ -77,7 +77,9 @@ test('salesforce helper never touches secrets directly', () => {
         'assert "subprocess" not in content, (',
         '    "Script must not shell out to manage secrets"',
         ')',
-        'assert "captureResponseSecrets" not in content or True, "sanity"',
+        'assert "captureResponseSecrets" not in content, (',
+        '    "Script must not reference captureResponseSecrets directly"',
+        ')',
         'print("ok")',
       ].join('\n'),
       helperPath,
